Ensure we walk the whole tree replacing references
Previously, we were only going one level deep.
When we replaced a reference to a definition,
we never checked to see if that def also referenced another definition.

My original attempt at this tried to process the definitions by finding
any references to them, and replacing it, removing the definition and
moving on, but that algorithm ended up being O(n^n).

The current algorithm does perform a lookup to see if any of the current
list of definitions are referenced, and we do recurse until none are
left, so this is definitely a brute force algorithm.
(I think it's O(n^2) or O(n^3), but I didn't take the time to actually analyze
it.)

There's room for improvement, but it does complete before the heat death
of the universe.